Database refactoring for packaging
