+++ src/main/java/constants/DatabaseConstants.java
@@ -1,9 +1,6 @@
 package constants;
 
 public class DatabaseConstants {
-
-
-
     public static final String DB_URL = "jdbc:mysql://aa1cma7c5gq1xq9.cjvinftl7gsm.us-east-2.rds.amazonaws.com:3306";
     public static final String DB_USERNAME = "admin";
     public static final String DB_PASSWORD = "WinnieThePooh!";
